Move isMlRule and RuleType to common
These are used on both the frontend and the backend, and can be shared.
