+++ x-pack/legacy/plugins/siem/public/containers/detection_engine/rules/use_rule_status.test.tsx
@@ -12,7 +12,7 @@ import {
   ReturnRulesStatuses,
 } from './use_rule_status';
 import * as api from './api';
-import { RuleType, Rule } from '../rules/types';
+import { Rule } from '../rules/types';
 
 jest.mock('./api');
 
@@ -57,7 +57,7 @@ const testRule: Rule = {
   threat: [],
   throttle: null,
   to: 'now',
-  type: 'query' as RuleType,
+  type: 'query',
   updated_at: 'mm/dd/yyyyTHH:MM:sssz',
   updated_by: 'mockUser',
 };
