+++ x-pack/legacy/plugins/siem/public/pages/detection_engine/rules/helpers.tsx
@@ -10,10 +10,11 @@ import moment from 'moment';
 import memoizeOne from 'memoize-one';
 import { useLocation } from 'react-router-dom';
 
-import { RuleAlertAction } from '../../../../common/detection_engine/types';
+import { RuleAlertAction, RuleType } from '../../../../common/detection_engine/types';
+import { isMlRule } from '../../../../common/detection_engine/ml_helpers';
 import { transformRuleToAlertAction } from '../../../../common/detection_engine/transform_actions';
 import { Filter } from '../../../../../../../../src/plugins/data/public';
-import { Rule, RuleType } from '../../../containers/detection_engine/rules';
+import { Rule } from '../../../containers/detection_engine/rules';
 import { FormData, FormHook, FormSchema } from '../../../shared_imports';
 import {
   AboutStepRule,
@@ -213,8 +214,6 @@ export const setFieldValue = (
       form.setFieldValue(key, val);
     }
   });
-
-export const isMlRule = (ruleType: RuleType) => ruleType === 'machine_learning';
 
 export const redirectToDetections = (
   isSignalIndexExists: boolean | null,
