+++ x-pack/legacy/plugins/siem/server/lib/detection_engine/types.ts
@@ -8,7 +8,7 @@ import { CallAPIOptions } from '../../../../../../../src/core/server';
 import { Filter } from '../../../../../../../src/plugins/data/server';
 import { IRuleStatusAttributes } from './rules/types';
 import { ListsDefaultArraySchema } from './routes/schemas/types/lists_default_array';
-import { RuleAlertAction } from '../../../common/detection_engine/types';
+import { RuleAlertAction, RuleType } from '../../../common/detection_engine/types';
 
 export type PartialFilter = Partial<Filter>;
 
@@ -28,7 +28,6 @@ export interface ThreatParams {
 // TODO: Eventually this whole RuleAlertParams will be replaced with io-ts. For now we can slowly strangle it out and reduce duplicate types
 // We don't have the input types defined through io-ts just yet but as we being introducing types from there we will more and more remove
 // types and share them between input and output schema but have an input Rule Schema and an output Rule Schema.
-export type RuleType = 'query' | 'saved_query' | 'machine_learning';
 
 export interface RuleAlertParams {
   actions: RuleAlertAction[];
